add new entry to db

+++ app/src/main/java/org/devcloud/snippets/app/DatabaseHelper.java
@@ -37,14 +37,26 @@ class DatabaseHelper extends SQLiteOpenHelper {
     db.execSQL("CREATE TABLE " + Snippet.TABLE_NAME + " (_id INTEGER PRIMARY KEY," +
         Snippet.COLUMN_NAME_TEXT + " " + Snippet.COLUMN_TYPE_TEXT + " NOT NULL, " +
         Snippet.COLUMN_NAME_DATE + " " + Snippet.COLUMN_TYPE_DATE + " NOT NULL" +
+        Snippet.COLUMN_NAME_USERID + " " + Snippet.COLUMN_TYPE_USERID + " NOT NULL" +
         ")");
   }
 
   public void onUpgrade(SQLiteDatabase db, int oldVersion, int newVersion) {
     // This database is only a cache for online data, so its upgrade policy is
     // to simply to discard the data and start over
-    db.execSQL("DROP TABLE IF EXISTS " + Snippet.TABLE_NAME);
-    onCreate(db);
+    switch (oldVersion) {
+      case 1: {
+        db.execSQL("DROP TABLE IF EXISTS " + Snippet.TABLE_NAME);
+        onCreate(db);
+      }
+      case 2:
+        if (newVersion < 2) {
+          db.execSQL("DROP TABLE IF EXISTS " + Snippet.TABLE_NAME);
+          onCreate(db);
+        } else {
+          db.execSQL("ALTER TABLE "  + Snippet.TABLE_NAME + " ADD COLUMN '" + Snippet.COLUMN_NAME_USERID + "' " + Snippet.COLUMN_TYPE_USERID);
+        }
+    }
   }
 
   public void onDowngrade(SQLiteDatabase db, int oldVersion, int newVersion) {
